fix: attach last output to process object
This fixes a bug where `git-sync--last-output` state was persisting
through runs of `git-sync--execute`. When `git-sync--execute` is where
several times for the current project, if there were no changes sincce the
last time git-sync was ran then when running `git-sync--has-changes-p`
will not print any output to stdout, this leads to
`git-sync--process-filter` not running. Then, when
`git-sync--process-sentinel` runs it will resolve the promise with
git-sync--last-output, but because `process-filter` didn't set it to
anything, it still contains the last value that was set in the
previous run.

By putting the value directly on the object we're able to ensure a
blank output state every invocation of `git-sync--execute-command`.